.gitignore: Remove .vscode from the list of ignored files The changes were done to remove the .vscode directory from the list of ignored files in the .gitignore file. This is because the .vscode directory contains important configuration files for the Visual Studio Code editor.
.vscode/launch.json: Add launch configurations for development
The changes were done to add launch configurations for development in the .vscode/launch.json file. These configurations allow for easy debugging and running of the project in different environments.

.vscode/tasks.json: Add build task
The changes were done to add a build task in the .vscode/tasks.json file. This task allows for easy building of the project using the npm build script.